Update login process to use environment variables for Supabase credentials
