+++ src/app/page.tsx
@@ -1,17 +1,25 @@
 import { createClient } from "@/utils/supabase/server";
 import { cookies } from "next/headers";
 
-import ClientComponentInitiatingJotai from "./ClientComponentInitiatingJotai";
 import ParentProvider from "./ParentProvider";
+// import ParentProvider from "./ParentProvider";
 
 import { useServerAuthSession } from "@/utils/authSession";
+import Stables from "./Stables";
 
 const App = async () => {
   const cookieStore = cookies();
   const supabase = createClient(cookieStore);
   // const { data } = await supabase.auth.getSession();
 
-  let { data: supbaseData, error } = await supabase.from("test").select("*");
+  // let { data: supbaseData, error } = await supabase.from("test").select("*");
+
+  let { data: stables, error: stablesError } = await supabase
+    .from("stables")
+    .select(
+      "id, name, admin_id, users!users_stables(id,first_name,last_name), feeders(id, name, stable_id), horses(id, name)",
+    )
+    .order("name");
 
   // if (!data.session) {
   //   return (
@@ -26,9 +34,9 @@ const App = async () => {
   return (
     <main className="flex min-h-screen items-center justify-center bg-neutral-900">
       <h1 className="font-roboto text-2xl font-bold text-neutral-200">
-        <ClientComponentInitiatingJotai data={supbaseData ?? []}>
-          <ParentProvider />
-        </ClientComponentInitiatingJotai>
+        <ParentProvider data={stables ?? []}>
+          <Stables />
+        </ParentProvider>
       </h1>
     </main>
   );
