+++ src/components/Login.tsx
@@ -32,8 +32,8 @@ const Login = () => {
     const supabase = createClient();
     // Sign in with email and password
     let { data: userData, error } = await supabase.auth.signInWithPassword({
-      email: "***REMOVED***",
-      password: "***REMOVED***",
+      email: process.env.NEXT_PUBLIC_SUPABASE_USERNAME!,
+      password: process.env.NEXT_PUBLIC_SUPABASE_PASSWORD!,
     });
     // userData && setUser(userData);
